filtering expanded tiles + minor changes

+++ cell2location/models/_cell2location_WTA_module.py
@@ -247,7 +247,7 @@ class LocationModelWTAMultiExperimentHierarchicalGeneLevel(PyroModule):
     def forward(self, x_data, neg_data, n_nuclei, idx, batch_index):
         self.n_neg_probes = neg_data.shape[1]
 
-        obs2sample = one_hot(batch_index, self.n_batch)
+        obs2sample = one_hot(batch_index, self.n_batch).float()
 
         obs_plate = self.create_plates(x_data, neg_data, n_nuclei, idx, batch_index)
